+++ frontend/src/modules/SubsurfaceMap/components/AggregationSelector/aggregationSelector.tsx
