+++ frontend/src/modules/SubsurfaceMap/components/SyncedSubsurfaceViewer/index.ts
