+++ frontend/src/modules/SubsurfaceMap/components/SyncedSubsurfaceViewer/syncedSubsurfaceViewer.tsx
