+++ frontend/src/modules/SubsurfaceMap/polygonsDirectoryProvider.ts
